requirements.txt
google-generativeai
python-docx
pypdf
PyPDF2
python-dotenv
tqdm
nltk
langchain>=0.1.0
langchain-community>=0.0.16
langchain-core>=0.1.10
langchain-google-genai>=0.0.8
langchain-pinecone>=0.0.3
langchain-text-splitters>=0.0.1
langchain-weaviate
unstructured>=0.10.30
passlib
weaviate-client

